rmv/add: move mainpage components under 'pages' directory
