+++ src/pages/mainpage/components/ItemList/ItemListBody.tsx
@@ -2,7 +2,7 @@
 import { css } from "@emotion/react";
 import { Col, Row } from "antd";
 import ItemButton from "./ItemListBody/ItemButton";
-import { ProductThumbnailInfo } from "../../../models/product";
+import { ProductThumbnailInfo } from "../../../../models/product";
 import { Typography } from "antd";
 
 const { Text, Title } = Typography;
