+++ src/pages/mainpage/components/ItemList/ItemListHeader.tsx
@@ -1,7 +1,7 @@
 /** @jsxImportSource @emotion/react */
 import { css } from "@emotion/react";
 import { Link } from "react-router-dom";
-import { colors } from "../../../styles/colors";
+import { colors } from "../../../../styles/colors";
 import { Typography, Space } from "antd";
 
 const { Title, Text } = Typography;
