+++ src/pages/mainpage/components/SearchSection.tsx
