+++ src/pages/mainpage/components/SearchSection/DepartmentDropdown.tsx
