+++ src/pages/mainpage/components/SearchSection/SearchInput.tsx
